Add everything required for systemd-sysupdate
Let's set things up for systemd-sysupdate. This means A/B partitioning
setup by repart on first boot, configuring the mkosi repart files to
give recognizable labels with the image version included and proper
split artifact names so the partition UUID to use can be extracted by
systemd-sysupdate.

We also add a set of systemd-sysupdate transfer definitions files in
mkosi.sysupdate to use with the upcoming mkosi sysupdate verb. These
can be used to build a new image with mkosi which is then deployed with
mkosi sysupdate.

To keep things simple, we make the image version the current date + time
so that it automatically always increases on each image build.