Use confext instead of tmpfiles to populate /etc
A confext is much easier to update compared to /etc populated at
first boot with tmpfiles. We'll add some features to mkosi to make
it trivially easy to rebuild just the confext so it can be very
quickly be rebuilt to play around with changes to /etc without having
to rebuild and reinstall the full image.