Store /etc in /usr and make /etc a symlink instead of using confext
confext has the issue that it can only be used to provide config for
services that are started after systemd-confext.service. Since our
goal is to have an image-based system that can be trivially and quickly
rebuilt by the user, let's instead put /etc in /usr and make /etc a
symlink to /usr.

Quick /etc config changes can still be made using systemd-confext
without needing a reboot. Persistent config changes will require
rebuilding the image with the config change and soft-rebooting.