Make image self replicating
Let's allow replicating the image to another device by adding CopyBlocks=auto
to the usr, usr-verity and esp partitions. We don't add it to the usr-verity-sig
partition as CopyBlocks=auto is not supported for usr-verity-sig partitions yet.
This isn't a problem as we don't rely on the usr-verity-sig partition yet as the
usrhash= is included in the kernel command line instead.