repart: include image name in home/swap/root
Let's associate these directories with our image. In a world where these
partitions are TPM locked it really matters we unlock them with the
right OS, because unlocking is unlikely to be possible on other images
since their PCRs won't match.

This is just preparation for an unknown, bright future. I have nothing
in petto that would actually care about the names at this point.