image-policy: tighten policy, insist on signing
Let's make the image policy stricter: we want to require signed veritfy,
not just plain verity logic when we dissect the root fs.

Note that currently this part of the image policy is not enforced either
way, but this is going to change soon (I hope).

Follow-up for: 03569f9f85ea45cf1176b6e3d78916ffffcc8e73