Rename mkosi.version to mkosi.bump
The History= option in mkosi only saves CLI arguments and not the entire
configuration anymore, meaning we can't rely on it anymore to save the
image version from mkosi.version.

Instead, we can use mkosi.bump which causes the new version to be written
to mkosi.version so that it's still saved and reused.

We also bump the minimum mkosi version to the latest git commit at the same
time.