desktop: Make sure NetworkManager doesn't mess around with transient hostname
Currently, the system comes up with a "particle-xxx" hostname, which then switches
to the hostname advertised by DHCP (if any) at some point in time. It turns out that
the hostname leaks into the user session via the $SESSION_MANAGER environment variable,
required for session management under X which is still used by e.g. ksmserver in KDE. If
you get lucky, the transient hostname will be changed after $SESSION_MANAGER has been
initialized for the session, which results in applications such as firefox that link to
libxau which does something with $SESSION_MANAGER to try and resolve the old transient
hostname which is not resolveable anymore, leading to severe slowdowns in application
startup.

While there's many layers to this problem that should be fixed, the only one we have
direct control over is whether NetworkManager manages the transient hostname or not, so
let's disable that feature for desktop systems at least, since it's very dubious that users
would want their system to pick up a transient hostname from DHCP in the first place there.