particleos: add support for sysinstall
This adds a new UKI profile for booting in installer mode, using the
recently merged sysinstall feature of systemd:

https://github.com/systemd/systemd/pull/41877

This also changes the live mode (which is very similar) to not disable
firstboot, but instead uses credentials to tweak the prompts. This is
beneficial, since we want firstboot to run always, not because of the
interactivity features, but simply because it processes a bunch of
firstboot.* credentials we want.

It also gets rid of the changes made to console logging. That's useful
for the debug boot mode, but is inappropriate for the installer and live
profiles which should just be "clean"

This drops CopyFiles= from the installed repart.d/ drop-in for the ESP:
we don't need it anymore, systemd-sysinstall's "bootctl link" and
"bootctl install" invocations are much better at setting up the ESP.