desktop: mask systemd-networkd-wait-online
Turns out presetting this to disable is impossible, and it keeps
coming back like a zombie. Even after removing all presets (from
particleos, mkosi and systemd repos, yes all 3 of them preset it to
enabled), it is still enabled by default. Mask it until it is
fixed to cohabitate with NetworkManager-wait-online.