fedora gnome: add support for gpg "pinentry"
GPG support matters, and if we install no "pinentry" implementation
entering an GPG passphrase won't work, hence let's install GNOME's
version of it on GNOME, which hooks this up to gnome shell.

(this is split out on on debian + fedora, hence this covers both. on
arch the thing is apparently part of the main pinentry package, hence
don't bother)